requirements.txt
bitstring
black
build
canopen
dacite
isort
pylama[all]
pylama[toml]
pyyaml
types-pyyaml
setuptools
setuptools-scm
sphinx
sphinx-rtd-theme
tabulate
types-tabulate
wheel
yamllint
